refactor: useModal hook 추가
