+++ src/hooks/useModal.js
@@ -1,0 +1,17 @@
+import { useState } from "react";
+
+const useModal = () => {
+  const [modalVisible, setModalVisible] = useState(false);
+
+  const openModal = () => {
+    setModalVisible(true);
+  };
+
+  const closeModal = () => {
+    setModalVisible(false);
+  };
+
+  return { modalVisible, openModal, closeModal };
+};
+
+export default useModal;
